fix(hnsw): use negative of the radius when the distance metric is dot product for HNSW range search
The `range_search` method in `IndexHNSW` uses a 'negative distance computer' to compare dot product
distances to maintain parity with L2 distance comparison. Essentially, the inequality 'distance >
threshold' is changed to 'threshold > -distance' instead of the correct '-threshold > -distance'.
So we multiply the radius (or threshold) by -1 to correct the inequality check.

add a descriptive comment