fix(vector_index): fix memory leak errors and OOM handling in Faiss
Summary:
Faiss has some memory allocations that potentially leak memory by using the
keyword `new` to allocate memory and either not having a corresponding `delete`
or missing/incorrect logic in destructors. Many of these were caught by stackhasher
and most of them occur in the index read code path that was identified in this diff:
https://grizzly.internal.memcompute.com/D66219.

Also added some asserts that fail in code that is currently not being used (index types
that memsql doesn't use but faiss supports) to identify and fix them if/when we add support
for them.

**Design doc/spec**:
**Docs impact**:
**Preliminary Reviewer(s)**:
**Final Reviewer**:

Test Plan: Pending

Reviewers: chunwu, szupo, pvetere

Reviewed By: chunwu, pvetere

Subscribers: jsherwood, engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D69909