fix(memory): use unique_ptr in get_CodePacker()
Summary:
The function get_CodePacker() is called to construct a new `CodePacker`
object which is passed to the constructor of `BlockInvertedLists`.
However, if allocation within the constructor fails (while resizing
`ids` or `codes` for example), the memory allocated to the `CodePacker`
object is never freed. So we pass in a `unique_ptr` whose ownership is
released at the end of the constructor. Note that the `packer` field
within `BlockInvertedLists` is deleted in its destructor.

**Design doc/spec**:
**Docs impact**: none
**Preliminary Reviewer(s)**:
**Final Reviewer**:

Test Plan:
memsql tests which use the updated faiss code
.psyduck: https://psyduck.internal.memcompute.com/run/08c53c8e956b4ee09abfabc4a834530c/tests/all
.vector_index: https://psyduck.internal.memcompute.com/run/d78d5899fb9d4344903889f93b2920c4/tests/all
vector_index/sanity.py with stackhash enabled:
https://fastduck.internal.memcompute.com/run/e1bdad23ef924dd19d5f1a032d0a9e95/tests/all

Reviewers: pvetere, ksaurav

Reviewed By: ksaurav

Subscribers: engineering-list

Differential Revision: https://grizzly.internal.memcompute.com/D73027